WIP
The real issue here is with composite charts. We need to be able to
pass in a group at "runtime" which basically means before you render.

Seems like the previous implementation where the base group was stored
off the groupStack was rather clever in that it made this behavior near
trivial to implement.

Unsure what the best approach is going forward with this refactoring.
Maybe delete this branch?